requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
numpy==1.19.2
numpy-financial==1.0.0
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
